requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.6.2.post1
blinker==1.9.0
certifi==2024.8.30
cffi==1.17.1
click==8.1.7
eval_type_backport==0.2.0
Flask==3.0.2
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
hume==0.7.4
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
load-dotenv==0.1.0
MarkupSafe==3.0.2
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
python-dotenv==1.0.1
simpleaudio==1.0.4
sniffio==1.3.1
sounddevice==0.4.7
typing_extensions==4.12.2
websockets==12.0
Werkzeug==3.0.1
